cmp wrapper used and checked on setWeeklyShifts - data fetching & cmp wrapper - done!
