+++ app/main/shifts-application/shifts_apply_page.tsx
@@ -1,15 +1,15 @@
 import { getEmployerWeeklyWorkflow } from '@/services/server-services/shifts.service';
 import WithDataWrapper from '@/components/helpers/cmp-wrapper';
-import { ShiftsTable } from './shifts-table';
+import { ShiftsTable } from './shifts-apply-table';
 import { Employee } from '@/types/class.service';
 
-export default function ShiftsApplyTable({ user }: { user: Employee }) {
+export default function ShiftsApplyPage({ user }: { user: Employee }) {
   
   const ShiftsTableWithData = WithDataWrapper({
     dataPromise: () => getEmployerWeeklyWorkflow(user.employer.id),
-    Component: (props) => <ShiftsTable {...props} user={user}  />, // Passing user prop
-    errorMsg: 'Couldnt load shifts',
-    loadingMsg: 'Loading Shifts...'
+    Component: (props) => <ShiftsTable {...props} user={user}  />, 
+    errorMsg: 'Couldnt load employees applications',
+    loadingMsg: 'Loading Shifts requests...'
   });
 
   return <ShiftsTableWithData />;
